server/pyproject.toml
[project]
name = "server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.17.2",
    "asgiref>=3.11.0",
    "celery>=5.6.0",
    "fastapi>=0.127.1",
    "google-genai>=1.56.0",
    "mem0ai>=1.0.1",
    "numpy>=2.4.0",
    "passlib[bcrypt]>=1.7.4",
    "pillow>=12.0.0",
    "psycopg2-binary>=2.9.11",
    "pydantic-settings>=2.12.0",
    "pydantic[email]>=2.12.5",
    "pydicom>=3.0.1",
    "pymupdf>=1.26.7",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.21",
    "redis>=7.1.0",
    "sqlalchemy-utils>=0.42.1",
    "sqlmodel>=0.0.30",
    "sse-starlette>=3.1.1",
    "uvicorn>=0.40.0",
]
